Implementación msvc-receta, @Update @Delete

+++ msvc-SolicitudDocente/src/main/java/msvc_SolicitudDocente/msvc_SolicitudDocente/models/DetalleProductoSolicitud.java
@@ -15,5 +15,5 @@ public class DetalleProductoSolicitud {
 
     private Long    idProducto;
 
-    private Integer cantidadUnidadMedida;
+    private Float cantidadUnidadMedida;
 }
